add cas login

+++ app/controllers/user_sessions_controller.rb
@@ -1,4 +1,5 @@
 class UserSessionsController < ApplicationController
+  before_filter RubyCAS::Filter, :only => :cas_login
   before_filter :require_no_user, :only => [:create]
   before_filter :require_user, :only => :destroy
   
@@ -19,8 +20,22 @@ class UserSessionsController < ApplicationController
   end
   
   def destroy
-    current_user_session.destroy
-    flash[:notice] = "Logout successful!"
-    redirect_back_or_default login_url
+    if session[:cas_user]
+      session[:cas_user] = nil
+      redirect_to "http://auth.tadnet.org/logout"
+    else
+      current_user_session.destroy if current_user_session
+      flash[:notice] = "Logout successful!"
+      redirect_back_or_default login_url
+    end
+  end
+
+  def cas_login
+    if current_user
+      flash[:notice] = "Logged in with CAS"
+    else
+      flash[:warning] = "Problem logging in with CAS"
+    end
+    redirect_back_or_default '/'
   end
 end
